requirements.txt
# FastAPI and ASGI server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy==2.0.36
alembic==1.14.0
pymysql==1.1.1
cryptography==44.0.0

# Pydantic settings
pydantic==2.10.4
pydantic-settings==2.7.0

# File handling
python-magic==0.4.27
aiofiles==24.1.0
Pillow==11.0.0

# Logging
python-json-logger==2.0.7

# Utilities
python-slugify==8.0.4
